SoapClient.cs - Add flag for key storage location
For our server setup, each grouping of webservices runs in its own AppPool. In our setup the 'Identity' of the 'Process Model' for the App Pool settings is not one of the standard built-in accounts (like NETWORK SERVICE). We use an AD account that we created for tracking purposes. Because the CyberSource certs are typically stored in the 'Current User' store and not the 'Local Computer' store all calls to CyberSource fail for "System.Security.Cryptography.CryptographicException: An internal error occurred." The App Pool identity needs access to the keys, so we add the extra flag of 'X509KeyStorageFlags.MachineKeySet' and all is well!