Update NVPClient.cs to support the recent .p12 password changes
With the change made after February 28th where the password to access the .p12 file is now set at the Cybersource dashboard, and it does not use the default password (merchantId) anymore, the CreateDnsIdentity is expecting the MerchantId and not the password. This will cause an exception to be thrown when trying to run the transaction: 
string src = nVPTransactionProcessorClient.runTransaction(Hash2String(request));

I'm suggesting this fix to create the dns identity using the merchantId and not the password.